requirements.txt
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.28.0
plotly>=5.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
asyncio
aiofiles>=23.0.0
asyncio
aiofiles==23.2.1
logging
datetime
json
csv
typing
dataclasses
abc
joblib
